+++ com.unity.mobile.android-logcat/Editor/AndroidLogcatSettingsProvider.cs
@@ -17,6 +17,7 @@ namespace Unity.Android.Logcat
             public static GUIContent stactraceRegex = new GUIContent("Stacktrace Regex", "Configure regex used for resolving function address and library name");
             public static GUIContent requestIntervalMS = new GUIContent("Request Interval ms",
                 $"How often to request memory dump from the device? The minimum value is {AndroidLogcatSettings.kMinMemoryRequestIntervalMS} ms");
+            public static GUIContent maxExitedPackageToShow = new GUIContent("Max Exited Packages", "The maximum number of pacakges in pacakge selection which have exited.");
         }
 
         private AndroidLogcatRuntimeBase m_Runtime;
@@ -55,6 +56,12 @@ namespace Unity.Android.Logcat
                 EditorGUILayout.IntField(Styles.requestIntervalMS, settings.MemoryRequestIntervalMS);
             GUILayout.Space(20);
 
+            GUILayout.Space(20);
+            EditorGUILayout.LabelField("Packages", EditorStyles.boldLabel);
+
+            settings.MaxExitedPackagesToShow = EditorGUILayout.IntSlider(Styles.maxExitedPackageToShow, settings.MaxExitedPackagesToShow, 1, 100);
+
+            GUILayout.Space(20);
             EditorGUILayout.LabelField(Styles.stactraceRegex, EditorStyles.boldLabel);
             m_RegexList.OnGUI(150.0f);